refactor: expose WebhooksList and JsonSerializable from public types module
