+++ src/apify_client/_resource_clients/actor.py
@@ -35,7 +35,6 @@ if TYPE_CHECKING:
     from decimal import Decimal
     from logging import Logger
 
-    from apify_client._internal_types import WebhooksList
     from apify_client._literals import ActorJobStatus, ActorPermissionLevel, RunOrigin
     from apify_client._resource_clients import (
         ActorVersionClient,
@@ -53,7 +52,7 @@ if TYPE_CHECKING:
         WebhookCollectionClient,
         WebhookCollectionClientAsync,
     )
-    from apify_client.types import Timeout
+    from apify_client.types import Timeout, WebhooksList
 
 _PricingInfo = (
     PayPerEventActorPricingInfo
