+++ src/apify_client/_resource_clients/dataset.py
@@ -20,10 +20,9 @@ if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
     from datetime import timedelta
 
-    from apify_client._internal_types import JsonSerializable
     from apify_client._literals import GeneralAccess
     from apify_client.http_clients import HttpResponse
-    from apify_client.types import Timeout
+    from apify_client.types import JsonSerializable, Timeout
 
 
 @docs_group('Other')
