+++ src/apify_client/_resource_clients/task.py
@@ -19,7 +19,6 @@ from apify_client._utils import encode_webhooks_to_base64, response_to_dict, to_
 if TYPE_CHECKING:
     from datetime import timedelta
 
-    from apify_client._internal_types import WebhooksList
     from apify_client._literals import ActorJobStatus, RunOrigin
     from apify_client._resource_clients import (
         RunClient,
@@ -30,7 +29,7 @@ if TYPE_CHECKING:
         WebhookCollectionClientAsync,
     )
     from apify_client._typeddicts import TaskInputDict
-    from apify_client.types import Timeout
+    from apify_client.types import Timeout, WebhooksList
 
 
 @docs_group('Resource clients')
