+++ src/apify_client/http_clients/_base.py
@@ -25,8 +25,7 @@ from apify_client._utils import to_seconds
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator, Mapping
 
-    from apify_client._internal_types import JsonSerializable
-    from apify_client.types import Timeout
+    from apify_client.types import JsonSerializable, Timeout
 
 
 @docs_group('HTTP clients')
